judged question./bisai/rect

+++ 124/p230.cpp
@@ -3,16 +3,16 @@
 using namespace std;
 #define MAXN 1000000
 
-int n,m;
-int rank[MAXN]={0};
-int father[MAXN]={0};
+long long n,m;
+long long rank[MAXN]={0};
+long long father[MAXN]={0};
 struct na{
-    map<int,int> e;
-    inline void add_edge(int to,int value){
+    map<long long,long long> e;
+    inline void add_edge(long long to,long long value){
         e[to]=value;
     }
-    inline int operator[](const int num){
-        map<int,int>::iterator iter = this->e.find(num);
+    inline long long operator[](const long long num){
+        map<long long,long long>::iterator iter = this->e.find(num);
         if(iter!=this->e.end())
         {
             return iter->second;
@@ -22,11 +22,11 @@ struct na{
 };
 na node[MAXN];
 void init(){
-    for(int i=1;i<=n;++i){
+    for(long long i=1;i<=n;++i){
         father[i]=i;
     }
 }
-inline int getfather(int pos){
+inline long long getfather(long long pos){
     if(father[pos]==pos){
         return pos;
     }else{
@@ -35,8 +35,8 @@ inline int getfather(int pos){
     }
 }
 
-inline void merge(int a,int b){
-    int fa,fb;
+inline void merge(long long a,long long b){
+    long long fa,fb;
     fa = getfather(a);
     fb=  getfather(b);
     if(fa==fb){
@@ -56,22 +56,22 @@ inline void merge(int a,int b){
 int main(){
     cin>>n;
     init();
-    int tmp;
-    for(int i=1;i<=n;++i){
+    long long tmp;
+    for(long long i=1;i<=n;++i){
         cin>>rank[i];
         cin>>tmp;
         node[i].add_edge(tmp,1);
         node[tmp].add_edge(i,1);
     }
-    for(int i=1;i<=n;++i){
-        for(int j=1;j<=n;++j){
+    for(long long i=1;i<=n;++i){
+        for(long long j=1;j<=n;++j){
             if(node[i][j]==0){
                 merge(i,j);
             }
         }
     }
-    int max = -1;
-    for(int i=1;i<=n;++i){
+    long long max = -1;
+    for(long long i=1;i<=n;++i){
         if(rank[i]>max){
             max = rank[i];
         }
